+++ stable_baselines3/common/envs/identity_env.py
