+++ tests/test_cnn.py
@@ -7,7 +7,7 @@ import torch as th
 from gym import spaces
 
 from stable_baselines3 import A2C, DQN, PPO, SAC, TD3
-from stable_baselines3.common.identity_env import FakeImageEnv
+from stable_baselines3.common.envs import FakeImageEnv
 from stable_baselines3.common.preprocessing import is_image_space, is_image_space_channels_first
 from stable_baselines3.common.utils import zip_strict
 from stable_baselines3.common.vec_env import VecTransposeImage, is_vecenv_wrapped
